successful mongodb connection to test database
