+++ .gitignore
@@ -1,5 +1,8 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
+# user secrets
+dbconfig.json
+
 # dependencies
 /node_modules
 /client/node_modules
