+++ package.json
@@ -2,13 +2,14 @@
   "name": "camp-express",
   "version": "0.0.0",
   "description": "Web application for dynamic camp information editing",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node server.js",
     "build-client": "cd client && yarn build && cd ..",
     "install-client": "yarn && cd client && yarn && cd .."
   },
   "dependencies": {
-    "express": "^4.16.2"
+    "express": "^4.16.2",
+    "mongodb": "^3.0.0-rc0"
   }
 }
